Small fix
GetType animation trả về initType, do type bất thường có thể dẫn đến lỗi stopping animation chỉ dc render flickering
Xóa timer fireball

+++ PlayScene.h
@@ -66,7 +66,7 @@ public:
 	void SetIsPause();
 	
 	bool GetIsStop() { return stopTimer->IsRunning(); }
-	void SetIsStop(int timeSpan) {
+	void SetIsStop(float timeSpan) {
 		stopTimer->SetTimeSpan(timeSpan);
 		stopTimer->Start();
 	}
